ml/pyproject.toml
[project]
name = "greyeye-ml"
version = "0.1.0"
description = "GreyEye ML training, evaluation, and export pipeline"
requires-python = ">=3.11"

dependencies = [
    "torch>=2.4",
    "torchvision>=0.19",
    "ultralytics>=8.3",
    "albumentations>=1.4",
    "onnx>=1.17",
    "onnxruntime>=1.19",
    "numpy>=1.26",
    "Pillow>=10.4",
    "pydantic>=2.9",
    "timm>=1.0",
    "pyyaml>=6.0",
    "shared-contracts",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ml"]

[tool.uv.sources]
shared-contracts = { workspace = true }
